refactor(parser): split parser.rs into parser/ directory
Split 1072-line monolith into focused modules:
- parser/mod.rs: Parser struct, parse_file(), re-exports
- parser/types.rs: Chunk, CallSite, FunctionCalls, ParserError
- parser/chunk.rs: chunk extraction, signatures, doc comments
- parser/calls.rs: call extraction, parse_file_calls

No behavior changes. All 258 tests pass.
